requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ci_sdr==0.0.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
ctc_segmentation==1.7.4
Cython==3.0.10
decorator==5.1.1
Distance==0.1.3
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
emoji==2.12.1
espnet==202301
espnet-tts-frontend==0.0.3
exceptiongroup==1.2.1
executing==2.0.1
fast_bss_eval==0.1.3
filelock==3.14.0
fsspec==2024.5.0
g2p-en==2.1.0
h5py==3.11.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
importlib-metadata==4.13.0
inflect==7.0.0
intel-openmp==2021.4.0
ipython==8.18.1
jaconv==0.3.4
jamo==0.4.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kaldiio==2.18.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
nltk==3.8.1
num2words @ git+https://github.com/savoirfairelinux/num2words.git@3e39091d052829fc9e65c18176ce7b7ff6169772
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==24.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
playsound==1.3.0
pooch==1.8.1
prompt_toolkit==3.0.45
protobuf==3.20.1
pure-eval==0.2.2
pvporcupine==3.0.2
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypinyin==0.44.0
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
pytorch-wpe==0.0.1
pytweening==1.2.0
pyworld==0.3.4
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
scikit-learn==1.5.0
scipy==1.11.4
sentencepiece==0.2.0
simpleaudio==1.0.4
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.4
stack-data==0.6.3
stanza==1.7.0
sympy==1.12.1
tbb==2021.12.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.3.0
torch-complex==0.4.3
torchaudio==2.3.0
tqdm==4.66.4
traitlets==5.14.3
tuya-connector-python==0.1.2
typeguard==2.13.3
typing_extensions==4.12.0
ukrainian-accentor @ git+https://github.com/egorsmkv/ukrainian-accentor.git@5b7971c4e135e3ff3283336962e63fc0b1c80f4c
ukrainian-tts @ git+https://github.com/robinhad/ukrainian-tts.git@bef46202e7dcfc77dc40e321f6f15b02eaef69d8
ukrainian-word-stress==1.1.0
Unidecode==1.3.8
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
zipp==3.19.0
